itemモデルの作成とマイグレーション
