+++ app/models/item.rb
@@ -1,7 +1,2 @@
 class Item < ApplicationRecord
-  # validates :name, presence: true
-  # validates :description, presence: true
-  # validates :price, presence: true
-
-  # belongs_to :user
 end
